+++ .github/workflows/coverage.yml
@@ -1,15 +1,11 @@
 name: Coverage
 
 on:
-  workflow_run:
-    workflows: End-to-end on push
-    types: 
-      - completed
+  workflow_dispatch:
 
 jobs:
   coverage:
     name: Combined coverage report
-    if: ${{ github.ref == 'refs/heads/master' }}
     runs-on: ubuntu-latest
     timeout-minutes: 5
     steps:
@@ -32,7 +28,7 @@ jobs:
       - name: Download e2e tests coverage report
         uses: dawidd6/action-download-artifact@v2
         with:
-          workflow: e2e-push.yml
+          workflow: e2e.yml
           branch: master
           path: test-coverage/e2e
